requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
weasyprint
jinja2
authlib
PyJWT
httpx
itsdangerous
watchgod
starlette
requests
bs4
apscheduler
python-jose